requirements.txt
discord.py>=1.2.5
python-box==3.4.6
ruamel.yaml==0.16.5
psycopg2-binary